Fix total_income type

+++ main.py
@@ -140,7 +140,7 @@ def load_airbnb_csv(event, context=None):
             bigquery.SchemaField("express_transfer_fee", "INTEGER"),
             bigquery.SchemaField("cleaning_fee", "INTEGER"),
             bigquery.SchemaField("pet_fee", "INTEGER"),
-            bigquery.SchemaField("total_income", "NUMERIC"),
+            bigquery.SchemaField("total_income", "INTEGER"),
             bigquery.SchemaField("accommodation_tax", "NUMERIC"),
             bigquery.SchemaField("hosting_revenue_fiscal_year", "INTEGER"),
             bigquery.SchemaField("row_id", "STRING", "REQUIRED"),
